test(range): migrate legacy a11y
