+++ core/src/components/range/test/legacy/a11y/range.e2e.ts-snapshots/range-active-md-ltr-Mobile-Firefox-linux.png
