+++ core/src/components/range/test/legacy/a11y/range.e2e.ts-snapshots/range-focus-md-ltr-Mobile-Safari-linux.png
